Fix bugs
Previously some results were not being painted by cert after hitting
the search button in amazon. If you then hit refresh in the browser the
results did get painted. So, it seems like amazon uses asynchronous
loading that was happening after the content script was already run.

Now, the search loop is triggered by any changes to the dom, using
jQuery to bind ‘DOMSubtreeModified’ on the body tag. This lead to
search being called >100 times on each page load, so I added a rate
limiter, which seemed to reduce it to ~10 searches per page load.